// material-kit-react/src/store/types.ts
export enum UserActionTypes {
    GET_USERS = 'GET_USERS',
    GET_USERS_SUCCESS = 'GET_USERS_SUCCESS',
    USERS_ERROR = 'USERS_ERROR',

    DIRECT_TO_DOMINATION = 'DIRECT_TO_DOMINATION',
    DIRECT_DOWN = 'DIRECT_DOWN',
    LOG_OUT = 'LOG_OUT',

    GET_PUBLIC_KEY = 'GET_PUBLIC_KEY',
    GET_PUBLIC_KEY_SUCCESS = 'GET_PUBLIC_KEY_SUCCESS',
    GET_PUBLIC_KEY_ERROR = 'GET_PUBLIC_KEY_ERROR',


    APPLY_PERMISSION='APPLY_PERMISSION',
    APPLY_PERMISSION_SUCCESS='APPLY_PERMISSION_SUCCESS',
    APPLY_PERMISSION_ERROR='APPLY_PERMISSION_ERROR',

    GET_APPLY_PERMISSION_LIST='GET_APPLY_PERMISSION_LIST',
    GET_APPLY_PERMISSION_LIST_SUCCESS='GET_APPLY_PERMISSION_LIST_SUCCESS',
    GET_APPLY_PERMISSION_LIST_ERROR='GET_APPLY_PERMISSION_LIST_ERROR',


    APPLY_PERMISSION_OPERATE='APPLY_PERMISSION_OPERATE',
    APPLY_PERMISSION_OPERATE_SUCCESS='APPLY_PERMISSION_OPERATE_SUCCESS',
    APPLY_PERMISSION_OPERATE_ERROR='APPLY_PERMISSION_OPERATE_ERROR',
}